[#2] refactor: listPage 폴더 구조 변경
page/list/index.tsx -> components/list/index.tsx
컴포넌트 분리 작업 진행
이후 추가될 가능성이 있는 dynamic import를 고려하여 분리
