+++ src/components/common/marginBox.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const MarginBox = styled.div<{ margin: string }>`
+  height: ${props => props.margin};
+`;
+
+export default MarginBox;
